astro/src/code-example-repositories/fusionauth-quickstart-python-django-web-main/complete-application/requirements.txt
fusionauth-client==1.47.0
Authlib==1.2.0
python-dotenv==1.0.0
requests==2.31.0
Django==3.2.15
mozilla-django-oidc==2.0.0